requirements.txt
asgiref==3.6.0
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==18.9.0
importlib-metadata==6.6.0
Markdown==3.4.3
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
zipp==3.15.0
